Refactor : PostCard 공통 컴포넌트 분리 및 프로필 글 탭 재사용
